reformat code, moved repository from controller

+++ src/main/java/ru/alexandrov/currencychecker/dao/model/XBTUSDModel.java
@@ -20,7 +20,7 @@ public class XBTUSDModel {
     @Column(name = "prev_price")
     private float prevPrice;
     @Column(name = "delta")
-    private int delta;
+    private float delta;
     @Column(name = "last_price")
     private float lastPrice;
     @Column(name = "increased")
