+++ src/main/resources/db/migration/V1.1__Initial.sql
@@ -3,8 +3,8 @@ CREATE TABLE BTCUSD
     id  SERIAL NOT NULL PRIMARY KEY,
     timestamp VARCHAR(255),
     currency VARCHAR(255),
-    prev_price INTEGER,
-    delta INTEGER,
-    last_price INTEGER,
+    prev_price FLOAT,
+    delta FLOAT,
+    last_price FLOAT,
     increased INTEGER
 );
